QQuickComboBox: ensure we don't close popup on iOS
On iOS (and Android), we give focus to a control on touch release
(rather than on touch begin, which is normal on desktop). For a
QQuickCombobox, this means that we will both handle focus and open
the popup on touch release, and not focus on touch begin
and open popup on touch release, like on desktop. Because of this, we
need to be more careful about when we choose to close the popup as
well, otherwise it will close down before it gets a chance to show on iOS.

This patch will check why we lose focus on both the inner line edit
and the combobox itself, and based on this, choose whether we should
close the popup. Basically, if focus is just transferred between the
popup button and the line edit, we choose to keep the popup open.

Fixes: QTBUG-70161
Change-Id: Iec242c0b5570844868085480fdf96fc49d189b82
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>