Fix memory "leaks" in qmlcachegen
The CompilationUnit class owns the unit data. An exception was made in
bootstrap builds, where it was up to the caller to free the memory. This
lead to cases in qmlcachegen where we didn't free the memory. This is
best fixed by unifying the behavior.

This fixes the build when using an ASAN enabled build, as the runtime
aborts after calling qmlcachegen due to "leaks".

Change-Id: I8b55b4e302a9569a1d4e09eeb488c479368b50f0
Reviewed-by: Lars Knoll <lars.knoll@qt.io>