MultiPointHandler: eliminate "no points" warning with native gestures
In handlePointerEventImpl, there is the call
  d->centroid.reset(d->currentPoints);
with the expectation that currentPoints is not empty.  But we weren't
populating it in case of a native gesture.

It still ends up being empty at the end of the gesture, but it's
normal to return false from wantsPointerEvent() when there are no
eligible points.

Fixes: QTBUG-70083
Change-Id: I12ca6460a24d2eb6c44a639f83ce3ff17eb37613
Reviewed-by: Jan Arve Sæther <jan-arve.saether@qt.io>