Doc: Add the moduleheader information
Although the doc project is named QtQuickControls,
the repo is still called qtquickcontrols2 and the
exported headers are installed under
QT_INSTALL_PREFIX/include/QtQuickControls2. By default,
Qdoc tries to look for the headers in a directory that
matches the doc project name, which is not the right
include directory in this case.

Change-Id: I77889fe35c028a2fc903d49b571a564c57326c21
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>