If DragHandler is dragged within its margin, don't jump
Not being pressed inside the target is a necessary but not sufficient
reason to reset m_pressTargetPos to the center of the target.  The
intention was rather to make the target jump into position when the
parent was a different item: e.g. if a Slider has a DragHandler whose
target is the slider's knob, you can start dragging anywhere on the
whole Slider but you want the knob to jump to the cursor position when
the drag begins.

While we're at it, both branches of the if in onGrabChanged()
are checking that target() isn't null, so we can move that check out.

Fixes: QTBUG-74966
Change-Id: I05be11d27422b070d941b9e43d4e1157e071c3a5
Reviewed-by: Jan Arve Sæther <jan-arve.saether@qt.io>