Only create the imports array if importCount is greater than 0
Since it is possible that CompilationUnit::instantiate() might be called
more than once when the importCount is 0 then it should only create the
imports array when it is greater than 0. This prevents a memory leak due
to the recreation of this array each time it is called even though there
is no imports to assign.

Change-Id: I5d84b01de10bff2ca25248251e8337839e434bd5
Reviewed-by: Simon Hausmann <simon.hausmann@qt.io>