Update the QML Runtime icon
A new one has been drawn; use it on all platforms, not just on macOS.
It is now also used as the default window icon, to be more identifiable
when minimized, etc.  Of course it can be overridden by plugins that
can be loaded from qml.

[ChangeLog][QtQml][qml] The QML Runtime tool now has an updated
application icon and a default window icon.  QtQuick applications
can still use QWindow::setIcon() to override the window icon.

Task-number: QTBUG-70826
Task-number: QTBUG-74662
Change-Id: I8671d0c99f7f4283dbe2dc4c605abb560f7bf1a1
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>