Restore value bindings when disabling a Binding element
We previously only restored script bindings that were replaced by
a Binding. Now we handle both.

[ChangeLog] QML Binding elements now support restoring previous values
of the bound property when the binding is disabled. This will be the
default behavior in Qt 5.15. Reliance on the old behavior of only
restoring binding, not literal values results in a warning now.

Fixes: QTBUG-33444
Change-Id: I833403b0645c08eee486fbd4acf5d3c7de2ef73a
Reviewed-by: Michael Brasser <michael.brasser@live.com>