Move the check for SSE2 into the configure system
The JIT requires SSE2. Without it we can still use the interpreter. If
you compile with SSE2 enabled and then run on a machine without SSE2 you
will get a more interesting crash now. However, in this day and age we
don't have to expect random pre-Pentium4 CPUs out in the wild anymore.
If you want to use such a thing, you will probably build a special
version of Qt for it.

Change-Id: I14a71cb83876d2ce7fdad012c385d0d4389e7ddf
Reviewed-by: Simon Hausmann <simon.hausmann@qt.io>