Add snapMode to DragHandler
This changes snap behavior slightly. Basically, it does not snap
anymore if the target() item is an ancestor of the parentItem().

In addition, we add a property that enables users to change the behavior.
(SnapIfPressedOutsideTarget has the old behavior)

[ChangeLog][QtQuick][Event Handlers] Added DragHandler.snapMode which can
be used to configure under which conditions the dragged item is snapped
to be below the cursor. The default mode is SnapAuto. The old behavior
can be obtained through the SnapIfPressedOutsideTarget mode.

Fixes: QTBUG-75661
Change-Id: Ibc00e8fbe31b779f8e817af1505e76425467d27a
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>