Re-add cache for looking up attached properties objects
c018df5b4075ae962966d4df7653d476dab02840 removed an optimization in
qmlAttachedPropertiesObject. The attached properties ID was implicitly
cached as it was static in a method templated by the type the attached
properties were registered for. We don't actually need the ID for this,
as we can also cache the function pointer.

Fixes: QTBUG-75609
Change-Id: I15e728ce4bbb89062a84ee6f113b1188faa32b92
Reviewed-by: Simon Hausmann <simon.hausmann@qt.io>