Avoid assertion when having a ShaderEffect as a delegate
The assertion occurred when it initially tried to create an item from the
delegate because it needed to show the contents of the first item when the
combobox was collapsed.

So, in order to do that ComboBox would create an object from the delegate
by calling delegateModel->object(). Since the ShaderEffect was not ready
to be constructed yet (it didn't have a window, see
QQuickOpenGLShaderEffect::maybeUpdateShaders() ) it would call polish on
itself in order to try later. Shortly after, it called
delegateModel->release(). This didn't delete the object immediately (it
called deleteLater() on the object), but it removed the engine and context
it was associated with).

However, this left it in the list of QQuickWindowPrivate::itemsToPolish.
Due to this, QQuickWindow would call updatePolish on it shortly after (as
a response to an expose event in our case). But the ShaderEffect was still
not associated with an engine, so it would ultimately assert with no
engine.

The fix is to simply check if the ShaderEffect has an engine, and return
early if that happens.

Then when the dropdown opens, the ShaderEffect will get a polish() because
of QQuickItemPrivate::recursiveRefFromEffectItem(), and it will have both
an engine and a window and will properly initialize its shaders.

Fixes: QTBUG-67343
Change-Id: Ib8ceb92d0fafb9b958407a50db5daa7b45e1d82a
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>