Remove null pointer checks for "this" from QQmlContextData::resolvedUrl()
According to ISO/IEC 14882 §5.2.5/3 "If E1 has the type “pointer to class
X,” then the expression E1->E2 is converted to the equivalent form
(*(E1)).E2;". Thus, as QQmlContextData::resolvedUrl() is non-static method,
it cannot be called on a null pointer because it leads to undefined
behavior, and “this“ in QQmlContextData::resolvedUrl() cannot be a null
pointer. According to this document:
https://gcc.gnu.org/gcc-6/porting_to.html , starting from version 6, GCC
optimizations remove null pointer checks for "this", since "the this
pointer can never be null, which is guaranteed by the language rules."
Thus, on the one hand the “if (ctxt)“ check in
QQmlContextData::resolvedUrl() does nothing, on the other “if (engine“
check leads to undefined behavior if ctxt/this == nullptr.

Task-number: QTBUG-75983
Change-Id: Idfb1e26758d83223bb0845139d63e2e8e80dc714
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>