Tooling: Guard use of QRegularExpression
The Qt we are using might be compiled without regular expressions. The
epxression in qqmldebugserver.cpp is actually not needed as
QString::toInt() can do this check for us. The ones in
globalinspector.cpp are only for reformatting the name in a nicer way
for the selection highlight. We can skip those in case there are no
regular expressions.

Change-Id: I71e3f0314d8a68a41c3da2568207e311c6b41d1b
Reviewed-by: Simon Hausmann <simon.hausmann@qt.io>