Fix workflow for shutting down busy type loader threads
We need to set the shutdown flag right away in order to prevent further
thread events from getting scheduled. Those might lead to deadlocks. On
the loader thread's side, however, we need to process the existing
events before we process the shutdown flag, in order not to leak the
events. After we are done with this routine, the thread will finish by
itself. We don't need to wait on our mutex anymore. It's enough to wait
on the thread itself.

Task-number: QTBUG-75777
Change-Id: Id0afcb492d84617c78ff3bf94d0cc402aef2faa4
(cherry-picked from commit 0521b55f201299ee9f031f61d5bf8f2be18cf8c5)
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Igor Kushnir <igorkuo@gmail.com>