Replace remaining QLinkedLists with QVector
In both cases, the container holds pointers, so stability of
references can be ruled out as a cause for using linked lists. Both
containers are also only ever appended to and then consumed, so there
are no insertions in the middle that a linked list may speed up. Last,
it also cannot be the sheer size of the container, as QLinkedList has
3x the memory consumption of a vector.

We conclude that none of the things that make QLinkedList a container
of choice apply here, so we can use a QVector instead.

In QSGAbstractSoftwareRenderer, there was an accessor for the linked
list member, but it appears unused, so was removed.

Change-Id: I273f89c0d1267444019088371a5eb0b3b32a763c
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>
Reviewed-by: Lars Knoll <lars.knoll@qt.io>