Fix deployment of QtQuick for Android in tst_parserstress
We rely on the qmlimportscanner to find out what to deploy, but it
is limited to scanning .qml files.
This test creates the QML code from C++, so qmlimportscanner fails to
detect those dependencies.

Therefore, we add a dummy_imports.qml file that has the sole purpose
of giving qmlimportscanner which additional dependencies the test has.

Change-Id: I97d706396f70b30558bb3f0e2303d39c9f321a2d
Fixes: QTBUG-73572
Reviewed-by: Frederik Gladhorn <frederik.gladhorn@qt.io>