Fix distance field cache growing outside limit
Due to the way the area allocator is used, where one area spans
all the textures, it is possible to have the situation where part
of a glyph is allocated to a texture N and the rest to N+1, i.e.
the split happens inside the glyph. In a case like this, the
cache would grow the texture in question to accommodate the glyph,
which would in turn cause the height of the texture to be too
large.

For most systems, this would not happen, since a single texture would
be enough for a font, but when it does the texture creation may fail.

The fix is to reduce the height of the area allocated to each texture
by one row height, so that if the glyph does happen to hit the border
between two and the texture needs to extend, it will still not exceed
the actual maximum height.

[ChangeLog][Text] Fixed a bug when displaying many characters from the
same font on a system with a low maximum texture size.

Task-number: QTBUG-76528
Change-Id: I1e559aaf90552afe8531872264186daa5ee61939
Reviewed-by: Simon Hausmann <simon.hausmann@qt.io>