Fix possible crash when distance field cache spans multiple textures
The allocated area for each texture has to have its origin at (0,0)
otherwise the actual required height of the cache is too low. When
using the texture upload workaround, this would crash, and when not
it would lead to missing glyphs.

Note that this is because the texture may contain a bit of empty
space at the top, since the area allocator may allocate this area
to a glyph which actually belongs to the previous texture (we use
a single area allocator for the entire cache and some glyphs may
overlap the limit between two textures).

[ChangeLog][Text] Fixed missing glyphs and in some cases crashes
when large character sets were being used on system with a low
maximum texture size.

Task-number: QTBUG-76528
Change-Id: I710ebbdd2feba4567505393fb12f89b5dd8501f1
Reviewed-by: Simon Hausmann <simon.hausmann@qt.io>